factor: UserData 생성 리팩토링
